+++ main.go
@@ -17,7 +17,7 @@ func main() {
 
 	// Create application with options
 	err := wails.Run(&options.App{
-		Title:  "MyNewEditor",
+		Title:  "RM Data Workbench",
 		Width:  1280,
 		Height: 720,
 		AssetServer: &assetserver.Options{